Conflicts:
	plugins/diagramEditor/test/jetbrains.mps.lang.editor.diagram.tests/models/jetbrains/mps/lang/editor/diagram/tests@tests.mps

+++ languages/baseLanguage/tuples/languages/tuples/source_gen/jetbrains/mps/baseLanguage/tuples/typesystem/namedTuple_assignableTo_namedTuple_InequationReplacementRule.java
@@ -25,7 +25,7 @@ public class namedTuple_assignableTo_namedTuple_InequationReplacementRule extend
 
   public void processInequation(final SNode subtype, final SNode supertype, final EquationInfo equationInfo, final TypeCheckingContext typeCheckingContext, IsApplicable2Status status, final boolean inequalityIsWeak, final boolean inequalityIsLessThan) {
     if (SLinkOperations.getTarget(subtype, "classifier", false) == SLinkOperations.getTarget(supertype, "classifier", false)) {
-      if (!((int) ListSequence.fromList(SLinkOperations.getTargets(subtype, "parameter", true)).count() == (int) ListSequence.fromList(SLinkOperations.getTargets(supertype, "parameter", true)).count())) {
+      if (!(ListSequence.fromList(SLinkOperations.getTargets(subtype, "parameter", true)).count() == ListSequence.fromList(SLinkOperations.getTargets(supertype, "parameter", true)).count())) {
         MessageTarget errorTarget = new NodeMessageTarget();
         IErrorReporter _reporter_2309309498 = typeCheckingContext.reportTypeError(equationInfo.getNodeWithError(), "Parameter types counts don't match", "r:e119dbbd-3529-4067-8bad-6b9edd79d0b6(jetbrains.mps.baseLanguage.tuples.typesystem)", "1239968591797", null, errorTarget);
         HUtil.addAdditionalRuleIdsFromInfo(_reporter_2309309498, equationInfo);
@@ -64,7 +64,7 @@ public class namedTuple_assignableTo_namedTuple_InequationReplacementRule extend
   public boolean checkInequation(final SNode subtype, final SNode supertype, final EquationInfo equationInfo, IsApplicable2Status status, final boolean inequalityIsWeak, final boolean inequalityIsLessThan) {
     boolean result_14532009 = true;
     if (SLinkOperations.getTarget(subtype, "classifier", false) == SLinkOperations.getTarget(supertype, "classifier", false)) {
-      if (!((int) ListSequence.fromList(SLinkOperations.getTargets(subtype, "parameter", true)).count() == (int) ListSequence.fromList(SLinkOperations.getTargets(supertype, "parameter", true)).count())) {
+      if (!(ListSequence.fromList(SLinkOperations.getTargets(subtype, "parameter", true)).count() == ListSequence.fromList(SLinkOperations.getTargets(supertype, "parameter", true)).count())) {
         result_14532009 = false;
       }
       {
