+++ workbench/mps-workbench/source_gen/jetbrains/mps/ide/actions/CreateRootNode_ActionGroup.java
@@ -38,7 +38,7 @@ public class CreateRootNode_ActionGroup extends GeneratedActionGroup {
   public void doUpdate(AnActionEvent event) {
     try {
       List<SModel> models = MPSCommonDataKeys.MODELS.getData(event.getDataContext());
-      if ((int) ListSequence.fromList(models).count() == 1 && !(ListSequence.fromList(models).first() instanceof EditableSModel && !(ListSequence.fromList(models).first().isReadOnly()))) {
+      if (ListSequence.fromList(models).count() == 1 && !(ListSequence.fromList(models).first() instanceof EditableSModel && !(ListSequence.fromList(models).first().isReadOnly()))) {
         event.getPresentation().setVisible(false);
         return;
       }
